Add files via upload
These are the screenshots of our meeting with RP(Requirment Provider) on Zoom Workplace. The meeting was about of 30 minutes and in the screenshots it is shown that all the members of the meeting have participated in the communication and also delivered the their point of view related to the project.